requirements.txt
streamlit
torch
torchvision
pillow
numpy
diffusers
transformers   
